fix(manage): move formatFee to shared util package
