+++ apps/manage/src/app/views/cases/view/question-utils.js
@@ -291,17 +291,3 @@ function camelCaseToTitleCase(str) {
 function titleCase(str) {
 	return str.charAt(0).toUpperCase() + str.slice(1);
 }
-
-/**
- *  * Formats a fee value to two decimal places with commas.
- *  * Accepts numbers or numeric strings.
- *  * @param {number|string} fee
- *  * @returns {string}
- */
-export function formatFee(fee) {
-	if (fee === null || fee === undefined || fee === '') return '';
-
-	const num = Number(String(fee).replace(/,/g, ''));
-	if (isNaN(num)) return String(fee);
-	return num.toLocaleString('en-GB', { minimumFractionDigits: 2, maximumFractionDigits: 2 });
-}
