+++ packages/lib/forms/custom-components/fee-amount/question.js
@@ -1,6 +1,6 @@
 import { BOOLEAN_OPTIONS } from '@planning-inspectorate/dynamic-forms/src/components/boolean/question.js';
 import OptionsQuestion from '@planning-inspectorate/dynamic-forms/src/questions/options-question.js';
-import { formatFee } from 'crowndev-manage/src/app/views/cases/view/question-utils.js';
+import { formatFee } from '../../../util/numbers.js';
 
 export default class FeeAmountQuestion extends OptionsQuestion {
 	constructor({ title, question, fieldName, url, hint, validators, html, feeAmountInputFieldName, feeAmountQuestion }) {
